move events database into upcoming events app
