+++ busy_beaver/apps/poller/api/update_events.py
@@ -3,7 +3,9 @@ import logging
 from flask import request
 from flask.views import MethodView
 
-from busy_beaver.apps.events_database.task import start_sync_event_database_task
+from busy_beaver.apps.upcoming_events.event_database.task import (
+    start_sync_event_database_task,
+)
 from busy_beaver.common.decorators import admin_role_required
 from busy_beaver.toolbox import make_response
